CODE-RUB: Moving GoogleReCaptcha config model under GoogleReCaptcha broker model folder
