+++ ISL.Providers.Captcha.GoogleReCaptcha/Brokers/GoogleReCaptchaBroker/GoogleReCaptchaBroker.cs
@@ -2,7 +2,7 @@
 // Copyright (c) North East London ICB. All rights reserved.
 // ---------------------------------------------------------
 
-using ISL.Providers.Captcha.GoogleReCaptcha.Models.Brokers;
+using ISL.Providers.Captcha.GoogleReCaptcha.Models.Brokers.GoogleReCaptcha;
 using System;
 using System.Collections.Generic;
 using System.Net.Http;
