+++ ISL.Providers.Captcha.GoogleReCaptcha/Models/Brokers/GoogleReCaptcha/GoogleReCaptchaConfigurations.cs
@@ -2,7 +2,7 @@
 // Copyright (c) North East London ICB. All rights reserved.
 // ---------------------------------------------------------
 
-namespace ISL.Providers.Captcha.GoogleReCaptcha.Models.Brokers
+namespace ISL.Providers.Captcha.GoogleReCaptcha.Models.Brokers.GoogleReCaptcha
 {
     public class GoogleReCaptchaConfigurations
     {
